requirements.txt
streamlit

# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq

# Embeddings
sentence-transformers

# Vector DB
chromadb

# PDF processing
pypdf

# Environment variables
python-dotenv

# Optional (recommended dependencies)
tqdm
numpy
scikit-learn
